requirements.txt
aiogram==3.3.0
flask
psycopg2-binary
aiogram==3.3.0